Update Vending-machine-fsm.cs
This C# code demonstrates the concept of finite-state machines. It uses an enum to represent different states and methods to define the possible transitions between states. The main purpose of the code is to simulate a vending machine that dispenses a product when the user inserts a coin and selects a product.

https://pt.wikipedia.org/wiki/M%C3%A1quina_de_estados_finita